Quick-Sort in C
Type the following
C program in a text editor and save it in the folder C:\Code with the filename Quicksort.c
Compile and execute this program. A run of this program is given below:
Enter the number of items in the list, N (2 <= N <= 20): 8 
Enter the 8 integers I (0 <= I <= 30000) separated by white spaces:
80 70 60 50 40 30 20 10 
Sorted List: 10 20 30 40 50 60 70 80
Thank you.